requirements.txt
# local package
-e .

altair==5.3.0
arrow==1.3.0
asn1crypto==1.5.1
attrs==23.2.0
azure-core==1.30.1
azure-identity==1.16.0
azure-keyvault-secrets==4.8.0
azure-storage-blob==12.19.1
binaryornot==0.4.4
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cookiecutter==2.6.0
cryptography==42.0.5
et-xmlfile==1.1.0
filelock==3.13.3
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
idna==3.6
importlib-metadata==6.11.0
isodate==0.6.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
maricovault @ https://pranaytestv37702324687.blob.core.windows.net/azureml/Environment/azureml-private-packages/maricovault-1.0.0-py3-none-any.whl#sha256=e75bba301e7a32d416654bc427de4d63e24063419bf99903f6e478f2358a7566
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
msal==1.28.0
msal-extensions==1.1.0
numpy==1.26.4
openpyxl==3.1.2
packaging==23.2
pandas==2.2.1
Pillow==9.5.0
platformdirs==3.11.0
portalocker==2.8.2
protobuf==4.25.3
pyarrow==15.0.2
pycparser==2.22
pydeck==0.8.0
Pygments==2.17.2
PyJWT==2.8.0
Pympler==1.0.1
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.1
pytz-deprecation-shim==0.1.0.post0
pywin32==306
PyYAML==6.0.1
referencing==0.34.0
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
six==1.16.0
smmap==5.0.1
snowflake-connector-python==3.7.1
snowflake-sqlalchemy==1.5.1
sortedcontainers==2.4.0
SQLAlchemy==1.4.52
streamlit==1.25.0
tenacity==8.2.3
text-unidecode==1.3
toml==0.10.2
tomlkit==0.12.4
toolz==0.12.1
tornado==6.4
types-python-dateutil==2.9.0.20240316
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==4.3.1
urllib3==2.2.1
validators==0.24.0
watchdog==4.0.0
zipp==3.18.1

# external requirements
python-dotenv>=0.5.1

